Use libdeflate instead of zlib for compression / decompression
libdeflate is a functional replaced for zlib. It is data compatible
and offers up to a 4x speed improvement.

libdeflate has no stream mode meaning all the data for each zip block
must be written in a single call. This only affects the meta data which
must now be combined into a single in-memory block before compression.
Apart from that, the code changes are quite minimal and straightforward.