Remove support for incremental edits
Incremental edits were added to support authoring of Ptex files that
don't fit in memory. This is no longer an issue, and the upcoming
changes to support parallel writing assume the file fits in memory so
there is no purpose for this feature anymore. It is also likely that it
was never used, and if it were it has the possibility of generating
inefficient Ptex files if the file were not later consolidated.