Allow multi-threaded writing
With this change, PtexWriter::writeFace may be called in parallel.
Mipmap generation is now done per face during writeFace so that all the
face data compression can be parallelized.

All the compressed face data is collected in memory and held until the
writer is closed and/or finish is called, thus the memory use will
likely increase.