Handle arbitrarily large Ptex face textures
When the total on-disk size of a face is too large to be stored in a
FaceDataHeader, the size will
be stored separately in a separate large-face header. This header is a
list of the 64-bit sizes of
just the "large" faces for a given level, and immediately follows the
LevelDataHeader.