Rework hash table to allow lock-free read at all times (e.g. during table write/grow)
Previously, the table header and table were in different memory
locations and could get out of sync
during table updates. Thus, the table needed to be locked for read when
the table was being
updated. This was also a potential race condition which could result in
a failure to find an object
if the table mask was read as the table was being grown.

With this change, the table header and table are in the same memory
block, accessed by a single
pointer, and thus are always valid and always in sync.