Remove internal usage of incremental edit mode
Given that the incremental edits are no longer supported, we should
avoid the deprecated form of PtexWriter::edit, and use the new form that
drops the incremental param. We should also no longer report the result
of the Ptexture::hasEdits function in ptexinfo.cpp, as that function is
also deprecated and always returns false.