Prevent multiple opens on same file
Previously, the Ptex file was opened in PtexCache::get before the reader
was added to the cache. If multiple threads simultaneously requested the
same file, it's possible that each would open the file before noticing
that it was opened by another thread. This was by design but was not
optimal.

With this change, the reader is added to the cache before the file is
opened, and thus visible to other threads. Concurrent open attempts are
then handled in the reader using a mutex, and the file will only be
opened once.