Process Mru list immediately, if possible
Previously, when a texture was unref'd, it would be logged as recently used by adding it to the Mru
list. Then, when the Mru list was full, all of the logged files would be processed by one of the
threads, adding them to the end of the Lru list. Updating the Lru list requires a mutex, and the
purpose of the Mru list is to prevent multiple threads waiting on the mutex.

The problem with the previous design is that the processing can be delayed, perhaps indefinitely if
one or only a few textures are ever accessed. One effect of this is that the stats may get updated
late, or possibly never.

With this change, an attempt is made to process the Mru list each time a file is logged, regardless
of whether the Mru list is full.